fix: refactoring network providers
