+++ keygen-service/src/main/java/com/isaacandrade/keygeneratorservice/keygen/application/snowflake/infra/network/HostnameResolver.java
@@ -1,0 +1,5 @@
+package com.isaacandrade.keygeneratorservice.keygen.application.snowflake.infra.network;
+
+public interface HostnameResolver {
+    String resolve();
+}
