+++ keygen-service/src/main/java/com/isaacandrade/keygeneratorservice/keygen/application/snowflake/infra/network/NetworkInterfaceProvider.java
@@ -1,0 +1,9 @@
+package com.isaacandrade.keygeneratorservice.keygen.application.snowflake.infra.network;
+
+import java.net.NetworkInterface;
+import java.net.SocketException;
+import java.util.Enumeration;
+
+public interface NetworkInterfaceProvider {
+    Enumeration<NetworkInterface> getNetworkInterfaces() throws SocketException;
+}
