+++ keygen-service/src/main/java/com/isaacandrade/keygeneratorservice/keygen/application/snowflake/infra/network/NodeInfoProvider.java
@@ -1,4 +1,4 @@
-package com.isaacandrade.keygeneratorservice.keygen.application.snowflake.infra.node;
+package com.isaacandrade.keygeneratorservice.keygen.application.snowflake.infra.network;
 
 public interface NodeInfoProvider {
     long getDatacenterId();
